renaming files
test should never have been on git as it was a temporary test now main is the working "main" file again
